Fix/animation issues (#1)
* Change toggling of loading bool to just set true

Toggling `isLoading` does not make sense. If `onAppear` is called twice it would invert the bool and break the state.

This also fixes one of the animation bugs where the view spins back and forth or behaves weirdly.

* Add a value: to the ContainerActivityView animation

The animation now only listens for the isLoading value, and not anything else that may interfer with the animation, such as resizing of views.